requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
python-dotenv
pika
requests
openai